Extract connect from the surah container
